Take -a and -p from environment as default
Take the values for the flags `-a` and `-p` from environment variables
as the default, still allowing them to be overridden on the command
line.

Sometimes using `servedir`, you may always want to listen on all
interfaces but the default invocation of `servedir` in a Makefile recipe
does not specify `-a`. You can now set `SERVEDIR_ALL_INTERFACES=true` in
the environment and these pre-canned `servedir` invocations can work
with your defaults.