Remove globals when parsing flags
Remove the use of globals - `flag.CommandLine` and `os.Args` - when
parsing flags, and instead take the args as a parameter to `parseFlags`
and use a `flag.FlagSet` instead of the global `flag.CommandLine`.

This makes testing easier as the flag set does not interact with the
`go test` command line flags and we don't have to worry about redefining
flags if we call `parseFlags()` multiple times in a test.